refactor(prerender): extract domain-block and status-json into utils modules
Move new helpers introduced in the preceding fix commit to focused modules:
- utils/domain-block.js: computeDomainBlock() pure function (403 circuit breaker logic)
- utils/status-json.js: readSiteStatusJson() and uploadStatusSummaryToS3() (all S3 status.json I/O)
- utils/utils.js: normalizePathname() export (was private in handler.js, needed by status-json)

handler.js is now a thin orchestrator: imports and calls these helpers rather than defining them.
uploadStatusSummaryToS3 is re-exported from handler.js for backward compat with existing tests.
